feature/025 Linux a Windows
