+++ src/modules/medical-consultation/dtos/circumferences.dto.ts
@@ -1,0 +1,18 @@
+import { ApiProperty } from '@nestjs/swagger';
+//TODO crear mensajes de validaciones
+import { MConsultationMessagesEnum } from '../enums/m_consultation.enum';
+
+export class CircumferencesDto {
+  @ApiProperty()
+  c_relaxed_arm: number;
+
+  @ApiProperty()
+  c_contracted_arm: number;
+
+  @ApiProperty()
+  c_leg_max: number;
+
+  c_corrected_arm: number;
+
+  c_leg_arm: number;
+}
