+++ src/modules/medical-consultation/dtos/get-medical-consultation.dto.ts
@@ -1,0 +1,53 @@
+import { ApiProperty } from '@nestjs/swagger';
+import { Exclude } from 'class-transformer';
+
+export class GetMedialConsultationDto {
+  @Exclude()
+  _id: string;
+
+  @Exclude()
+  __v: number;
+
+  @ApiProperty()
+  user_uuid: string;
+
+  @ApiProperty()
+  patient_uuid: string;
+
+  @ApiProperty()
+  basic_measurements: Array<string>;
+
+  @ApiProperty()
+  body_measurements: Array<string>;
+
+  @ApiProperty()
+  vital_signs: Array<string>;
+
+  @ApiProperty()
+  plyometric_measurements: Array<string>;
+
+  @ApiProperty()
+  somatocarta: number;
+
+  @ApiProperty()
+  somatotype: number;
+
+  @ApiProperty()
+  fao_who_onu: number;
+
+  @ApiProperty()
+  harris_benedict: number;
+
+  @ApiProperty()
+  valencia: number;
+
+  @ApiProperty()
+  mifflin_st: number;
+
+  @ApiProperty()
+  readonly created_at: string;
+  @Exclude()
+  readonly updated_at: string;
+  @Exclude()
+  readonly deleted_at: string;
+}
