+++ src/modules/patients/errors/patient-not-found.error.ts
@@ -1,0 +1,7 @@
+import { PatientMessagesEnum } from '../enums/patients.enum';
+
+export class PatientNotFoundError extends Error {
+  constructor() {
+    super(PatientMessagesEnum.NOT_FOUND);
+  }
+}
