+++ src/modules/patients/static/index.ts
