+++ src/modules/patients/types/post-messages.type.ts
@@ -1,0 +1,6 @@
+import { CustomMessage } from 'src/shared/errors/custom.error';
+import { PatientMessagesEnum } from '../enums/patients.enum';
+
+export type PostMessagesType = {
+  [key in PatientMessagesEnum]: CustomMessage;
+};
