--- src/patients/errors/post-not-found.error.ts
@@ -1,7 +1,0 @@
-import { PostMessagesEnum } from '../enums/post.enum';
-
-export class PostNotFoundError extends Error {
-  constructor() {
-    super(PostMessagesEnum.NOT_FOUND);
-  }
-}
