+++ src/shared/interfaces/index.interfaces.ts
@@ -1,6 +1,5 @@
-import { QueryPatientDto } from "src/patients/dtos/query-patient.dto";
-import { PaginateDto } from "../dtos/paginate.dto";
-
+import { QueryPatientDto } from 'src/modules/patients/dtos/query-patient.dto';
+import { PaginateDto } from '../dtos/paginate.dto';
 
 export interface GetPaginatedPatient {
   paginate: PaginateDto;
